reorganize rust folder structure
