+++ src-tauri/src/lib.rs
@@ -2,19 +2,15 @@ use std::sync::{Arc, Mutex};
 
 use tauri::Manager;
 
-use crate::menus::menu_builders;
-
-mod file;
 mod menus;
 mod pages;
 mod project;
-mod states;
 mod track;
 mod types;
 
 #[cfg_attr(mobile, tauri::mobile_entry_point)]
 pub async fn run() {
-    let state_audio_context = states::StateAudioContext::new();
+    let state_audio_context = project::states::StateAudioContext::new();
     let master_output_device = state_audio_context
         .output_device()
         .expect("Failed to get master output device");
@@ -22,11 +18,11 @@ pub async fn run() {
     tauri::Builder::default()
         .manage(state_audio_context)
         .setup(move |app| {
-            let menu = menu_builders::build_menus(app);
+            let menu = menus::menu_builders::build_menus(app);
             app.set_menu(menu)?;
 
-            let state_mixer = states::StateMixer::new(&app.handle(), master_output_device.clone());
-            let state_mixer_guard = states::StateMixerGuard(Arc::new(Mutex::new(state_mixer)));
+            let state_mixer = project::states::StateMixer::new(&app.handle(), master_output_device.clone());
+            let state_mixer_guard = project::states::StateMixerGuard(Arc::new(Mutex::new(state_mixer)));
             app.handle().manage(state_mixer_guard);
 
             Ok(())
@@ -34,16 +30,16 @@ pub async fn run() {
         .on_menu_event(|app, event| {
             let app = app.clone();
             tauri::async_runtime::spawn(async move {
-                menu_builders::handle_menu_events(&app, &event).await;
+                menus::menu_builders::handle_menu_events(&app, &event).await;
             });
         })
         .plugin(tauri_plugin_opener::init())
         .plugin(tauri_plugin_dialog::init())
-        .invoke_handler(tauri::generate_handler![
-            track::get_track_list,
-            track::update_track,
-            menus::project_menu::add_empty_track
-        ])
+        .invoke_handler(tauri::generate_handler!(
+            track::commands::get_track_list,
+            track::commands::update_track,
+            menus::commands::add_empty_track
+        ))
         .run(tauri::generate_context!())
         .expect("error while running tauri application");
 }
