+++ src-tauri/src/menus/commands.rs
@@ -1,9 +1,9 @@
-use crate::{states, track};
+use crate::{project, track};
 
 #[tauri::command]
 pub fn add_empty_track(
-    mixer: tauri::State<states::StateMixerGuard>,
-    audio: tauri::State<states::StateAudioContext>,
+    mixer: tauri::State<project::states::StateMixerGuard>,
+    audio: tauri::State<project::states::StateAudioContext>,
     app: tauri::AppHandle,
 ) -> Result<(), String> {
     let state_mixer = mixer.0.lock().map_err(|_| "Failed to lock state mixer")?;
@@ -11,9 +11,9 @@ pub fn add_empty_track(
     let list = state_mixer.track_list.clone();
     let mut track_list = list.lock().map_err(|_| "Failed to lock track list")?;
     if let Some(input) = audio_context.input_device() {
-        let track = track::AudioTrack::new(
-            track::TrackType::In,
-            Some(track::StreamSource::new(&app, input)),
+        let track = track::track::AudioTrack::new(
+            track::track::TrackType::In,
+            Some(track::source::StreamSource::new(&app, input)),
             None,
         );
         let new_track_name = format!("track-{}", track_list.track_list().len() + 1);
