+++ src-tauri/src/project/mod.rs
@@ -1,0 +1,3 @@
+pub mod commands;
+pub mod file;
+pub mod states;
