+++ src-tauri/src/project/states.rs
@@ -1,21 +1,12 @@
-use std::{
-    collections::HashMap,
-    sync::{
-        atomic::{AtomicUsize, Ordering},
-        Arc, Mutex,
-    },
-};
+use std::{collections::HashMap, sync::{atomic::{AtomicUsize, Ordering}, Arc, Mutex}};
 
 use cpal::Device;
 
-use crate::{
-    track,
-    types::{InputDeviceRegistry, OutputDeviceRegistry},
-};
+use crate::{track, types};
 
 #[derive(bincode::Encode, bincode::Decode)]
 pub struct StateMixerRaw {
-    track_list: HashMap<String, track::AudioTrackRaw>,
+    track_list: HashMap<String, track::raw::AudioTrackRaw>,
 }
 
 impl From<StateMixer> for StateMixerRaw {
@@ -31,15 +22,15 @@ pub struct StateMixerGuard(pub Arc<Mutex<StateMixer>>);
 
 #[derive(Clone)]
 pub struct StateMixer {
-    pub track_list: Arc<Mutex<track::TrackList>>,
+    pub track_list: Arc<Mutex<track::track_list::TrackList>>,
 }
 
 impl StateMixer {
     pub fn new(app: &tauri::AppHandle, master_output: Arc<Device>) -> Self {
-        let mut track_list = track::TrackList::new();
-        let master_out = track::AudioTrack::new(
-            track::TrackType::MasterOut,
-            Some(track::StreamSource::new(app, master_output)),
+        let mut track_list = track::track_list::TrackList::new();
+        let master_out = track::track::AudioTrack::new(
+            track::track::TrackType::MasterOut,
+            Some(track::source::StreamSource::new(app, master_output)),
             None,
         );
         track_list.add_track("master-out", master_out);
@@ -53,15 +44,15 @@ impl From<StateMixerRaw> for StateMixer {
     fn from(value: StateMixerRaw) -> Self {
         let track_list = value.track_list;
         StateMixer {
-            track_list: Arc::new(Mutex::new(track::TrackList::from_raw(track_list))),
+            track_list: Arc::new(Mutex::new(track::track_list::TrackList::from_raw(track_list))),
         }
     }
 }
 
 #[derive(Clone)]
 pub struct StateAudioContext {
-    pub input_device_registry: Arc<InputDeviceRegistry>,
-    pub output_device_registry: Arc<OutputDeviceRegistry>,
+    pub input_device_registry: Arc<types::InputDeviceRegistry>,
+    pub output_device_registry: Arc<types::OutputDeviceRegistry>,
     pub input_device_index: Arc<AtomicUsize>,
     pub output_device_index: Arc<AtomicUsize>,
 }
@@ -69,8 +60,8 @@ pub struct StateAudioContext {
 impl StateAudioContext {
     pub fn new() -> Self {
         let host = cpal::default_host();
-        let input_device_registry = Arc::new(InputDeviceRegistry::new(&host));
-        let output_device_registry = Arc::new(OutputDeviceRegistry::new(&host));
+        let input_device_registry = Arc::new(types::InputDeviceRegistry::new(&host));
+        let output_device_registry = Arc::new(types::OutputDeviceRegistry::new(&host));
 
         StateAudioContext {
             input_device_registry,
